pyproject.toml
[tool.poetry]
name = "llmod"
version = "0.1.0"
description = ""
authors = ["Anton Solbjørg <4324290+Notnaton@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
discord-py = "^2.3.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
llmod = "llmod.main:main"
